refactor utils folder structure
