+++ components/SwappingModal/SwappingModal.tsx
@@ -7,8 +7,7 @@ import { FaArrowRight } from 'react-icons/fa';
 import { TokenDetails } from '@/interfaces';
 import { Button } from '@mui/material';
 import getTokenApproval from '@/utils/contract-methods/getTokenApproval';
-import { swapExactTokensForTokens } from '@/utils/swapV2exacttokensfortokens';
-import { swapV3 } from '@/utils/contract-methods/swapTokens';
+import { swapV3, swapExactTokensForTokens } from '@/utils/contract-methods/swapTokens';
 import CircularProgress from '@mui/material/CircularProgress';
 import addresses from "@/utils/address.json";
 import { tickToPrice } from '@/utils/utils';
